requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-auth-adfs==1.14.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
msal==1.28.1
psycopg==3.1.19
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.2
whitenoise==6.7.0
